# backend/item-service/docs/swagger.yaml
basePath: /api
definitions:
  models.AnyItem:
    properties:
      additional:
        additionalProperties: {}
        type: object
      address:
        type: string
      bought:
        type: boolean
      brand:
        type: string
      buy_url:
        type: string
      category:
        type: string
      color:
        type: string
      created_at:
        type: string
      cuisine_type:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      item_media_folder:
        type: string
      item_name:
        type: string
      item_type:
        type: string
      main_food:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      notes:
        type: string
      owner:
        type: string
      quantity:
        type: integer
      shop_name:
        type: string
      size:
        type: string
      sugar:
        type: string
      time_to_eat:
        type: string
      toppings:
        items:
          type: string
        type: array
      uuid:
        description: |-
          Generated by the frontend when the Add page opens, so the link parser
          channel and the saved item share one id. Null for older items.
        type: string
    type: object
host: localhost:8002
info:
  contact: {}
  description: Microservice for item and history management.
  title: EmIuMuaGi Item Service API
  version: "1.0"
paths:
  /items:
    get:
      description: Get a list of all items including clothes, food_and_drink, others,
        and restaurant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AnyItem'
            type: array
      security:
      - BearerAuth: []
      summary: List all items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Create a new item (clothes, food_and_drink, others, or restaurant)
      parameters:
      - description: Item to create
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.AnyItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AnyItem'
      security:
      - BearerAuth: []
      summary: Create an item
      tags:
      - items
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"