# backend/item-service/handlers/items.go
package handlers

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"log"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/viettungvuong/emiumuagi-backend/database"
	"github.com/viettungvuong/emiumuagi-backend/internal"
	"github.com/viettungvuong/emiumuagi-backend/models"
)

const NTFY_TOPIC = "emiumuagi_tung"

func sendPushNotification(title string, message string) {
	go func() {
		url := "https://ntfy.sh/" + NTFY_TOPIC
		req, err := http.NewRequest("POST", url, bytes.NewBuffer([]byte(message)))
		if err != nil {
			log.Println("Failed to create request:", err)
			return
		}
		req.Header.Set("Title", title)
		req.Header.Set("Tags", "shopping_bags,heart")

		client := &http.Client{Timeout: 3 * time.Second}
		resp, err := client.Do(req)
		if err != nil {
			log.Println("Failed to send notification:", err)
			return
		}
		log.Println("Notification sent about new item added")
		defer resp.Body.Close()
	}()
}

// GetItems retrieves all items with their specific type details
// @Summary List all items
// @Description Get a list of all items including clothes, food_and_drink, others, and restaurant
// @Tags items
// @Produce json
// @Security BearerAuth
// @Success 200 {array} models.AnyItem
// @Router /items [get]
func GetItems(c *gin.Context) {
	type PolledItem struct {
		models.Item
		CSize       *string `gorm:"column:c_size"`
		Color       *string `gorm:"column:color"`
		Brand       *string `gorm:"column:brand"`
		Sugar       *string `gorm:"column:sugar"`
		FSize       *string `gorm:"column:f_size"`
		FNotes      *string `gorm:"column:f_notes"`
		Toppings    *string `gorm:"column:toppings"`
		Category    *string `gorm:"column:category"`
		ONotes      *string `gorm:"column:o_notes"`
		MainFood    *string `gorm:"column:main_food"`
		CuisineType *string `gorm:"column:cuisine_type"`
		Address     *string `gorm:"column:address"`
		TimeToEat   *string `gorm:"column:time_to_eat"`
	}

	owners, err := internal.OwnerScope(c)
	if err != nil {
		log.Printf("Could not resolve partner, showing own items only: %v", err)
	}

	var results []PolledItem

	err = database.DB.Raw(`
		SELECT i.id, i.uuid, i.item_name, i.quantity, i.buy_url, i.shop_name, i.created_at, i.item_type, i.bought, i.owner,
			c.size as c_size, c.color, c.brand,
			f.sugar, f.size as f_size, f.notes as f_notes, f.toppings,
			o.category, o.notes as o_notes,
			r.main_food, r.cuisine_type, r.address, r.time_to_eat
		FROM items i
		LEFT JOIN clothes c ON i.id = c.id
		LEFT JOIN food_and_drinks f ON i.id = f.id
		LEFT JOIN others o ON i.id = o.id
		LEFT JOIN restaurants r ON i.id = r.id
		WHERE i.owner IN ? AND i.deleted_at IS NULL
		ORDER BY i.created_at DESC
	`, owners).Scan(&results).Error

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to retrieve items"})
		return
	}

	responses := make([]models.AnyItem, 0, len(results))
	for _, res := range results {
		resp := models.AnyItem{Item: res.Item}
		if res.ItemType == "clothes" {
			resp.Size = res.CSize
			resp.Color = res.Color
			resp.Brand = res.Brand
		} else if res.ItemType == "food_and_drink" {
			resp.Sugar = res.Sugar
			resp.Size = res.FSize
			resp.Notes = res.FNotes
			if res.Toppings != nil && *res.Toppings != "" && *res.Toppings != "null" {
				var t []string
				if json.Unmarshal([]byte(*res.Toppings), &t) == nil {
					resp.Toppings = t
				}
			}
		} else if res.ItemType == "others" {
			resp.Category = res.Category
			resp.Notes = res.ONotes
		} else if res.ItemType == "restaurant" {
			resp.MainFood = res.MainFood
			resp.CuisineType = res.CuisineType
			resp.Address = res.Address
			resp.TimeToEat = res.TimeToEat
		}
		responses = append(responses, resp)
	}

	c.JSON(http.StatusOK, responses)
}

func extractMetadata(item models.Item){
	
}

// CreateItem creates a new item
// @Summary Create an item
// @Description Create a new item (clothes, food_and_drink, others, or restaurant)
// @Tags items
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param item body models.AnyItem true "Item to create"
// @Success 201 {object} models.AnyItem
// @Router /items [post]
func CreateItem(c *gin.Context) {
	var input models.AnyItem
	if err := c.ShouldBindJSON(&input); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	if input.Quantity == 0 {
		input.Quantity = 1
	}

	tx := database.DB.Begin()

	item := input.Item
	// Automatically assign the authenticated user's username as the owner
	item.Owner = c.GetString("username")

	if err := tx.Create(&item).Error; err != nil {
		tx.Rollback()
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create base item"})
		return
	}

	if item.ItemType == "clothes" {
		cItem := models.Clothes{ID: item.ID, Size: input.Size, Color: input.Color, Brand: input.Brand}
		if err := tx.Create(&cItem).Error; err != nil {
			tx.Rollback()
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create clothes item"})
			return
		}
	} else if item.ItemType == "food_and_drink" {
		fItem := models.FoodAndDrink{ID: item.ID, Sugar: input.Sugar, Size: input.Size, Notes: input.Notes, Toppings: input.Toppings}
		if err := tx.Create(&fItem).Error; err != nil {
			tx.Rollback()
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create food_and_drink item"})
			return
		}
	} else if item.ItemType == "others" {
		oItem := models.Others{ID: item.ID, Category: input.Category, Notes: input.Notes}
		if err := tx.Create(&oItem).Error; err != nil {
			tx.Rollback()
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create others item"})
			return
		}
	} else if item.ItemType == "restaurant" {
		rItem := models.Restaurant{ID: item.ID, MainFood: input.MainFood, CuisineType: input.CuisineType, Address: input.Address, TimeToEat: input.TimeToEat}
		if err := tx.Create(&rItem).Error; err != nil {
			tx.Rollback()
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create restaurant item"})
			return
		}
	}

	tx.Commit()

	input.Item = item
	sendPushNotification("Mới mún mua thêm đồ nè! 🛍️", "Vừa thêm vào danh sách: "+item.ItemName)
	log.Printf("Added new item %s\n", item.ItemName)
	c.JSON(http.StatusCreated, input)
}

func DeleteItem(c *gin.Context) {
	id := c.Param("item_id")
	currentUser := c.GetString("username")
	var item models.Item

	if err := database.DB.Where("id = ? AND deleted_at IS NULL", id).First(&item).Error; err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Item not found"})
		return
	}

	// Only the person whose list it is can remove an item
	if item.Owner != currentUser {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "You are not allowed to delete this item"})
		return
	}

	// Soft delete: stamps DeletedAt and leaves the type-specific row alone, so
	// the nightly purge can clear both once the retention window has passed
	now := time.Now()
	if err := database.DB.Model(&item).Update("deleted_at", now).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete item"})
		return
	}

	log.Printf("Deleted %s\n", item.ItemName)

	c.Status(http.StatusNoContent)
}

func addHistory(ctx context.Context, item_id uint) uuid.UUID {
	h := models.History{
		ItemID: item_id,
		Time:   time.Now(),
	}

	// Store on db
	result := database.DB.WithContext(ctx).Create(&h)

	if result.Error != nil {
		return uuid.Nil
	}

	fmt.Printf("Add history for %+v successfully\n", item_id)

	return h.ID
}

func MarkItemAsBought(c *gin.Context) {
	id := c.Param("item_id")

	item, ok := loadItemInScope(c, id)
	if !ok {
		return
	}

	item.Bought = true
	database.DB.Save(&item)

	// Automatically create a history entry
	historyId := addHistory(c.Request.Context(), item.ID)

	var res struct {
		models.Item
		CSize       *string `gorm:"column:c_size"`
		Color       *string `gorm:"column:color"`
		Brand       *string `gorm:"column:brand"`
		Sugar       *string `gorm:"column:sugar"`
		FSize       *string `gorm:"column:f_size"`
		FNotes      *string `gorm:"column:f_notes"`
		Toppings    *string `gorm:"column:toppings"`
		Category    *string `gorm:"column:category"`
		ONotes      *string `gorm:"column:o_notes"`
		MainFood    *string `gorm:"column:main_food"`
		CuisineType *string `gorm:"column:cuisine_type"`
		Address     *string `gorm:"column:address"`
		TimeToEat   *string `gorm:"column:time_to_eat"`
	}

	database.DB.Raw(`
		SELECT i.id, i.uuid, i.item_name, i.quantity, i.buy_url, i.shop_name, i.created_at, i.item_type, i.bought, i.owner,
			c.size as c_size, c.color, c.brand,
			f.sugar, f.size as f_size, f.notes as f_notes, f.toppings,
			o.category, o.notes as o_notes,
			r.main_food, r.cuisine_type, r.address, r.time_to_eat
		FROM items i
		LEFT JOIN clothes c ON i.id = c.id
		LEFT JOIN food_and_drinks f ON i.id = f.id
		LEFT JOIN others o ON i.id = o.id
		LEFT JOIN restaurants r ON i.id = r.id
		WHERE i.id = ? AND i.deleted_at IS NULL
	`, id).Scan(&res)

	resp := models.AnyItem{Item: res.Item, Additional: map[string]any{
		"HistoryID": historyId,
	}}

	switch res.ItemType {
	case "clothes":
		resp.Size = res.CSize
		resp.Color = res.Color
		resp.Brand = res.Brand
	case "food_and_drink":
		resp.Sugar = res.Sugar
		resp.Size = res.FSize
		resp.Notes = res.FNotes
		if res.Toppings != nil && *res.Toppings != "" && *res.Toppings != "null" {
			var t []string
			if json.Unmarshal([]byte(*res.Toppings), &t) == nil {
				resp.Toppings = t
			}
		}
	case "others":
		resp.Category = res.Category
		resp.Notes = res.ONotes
	case "restaurant":
		resp.MainFood = res.MainFood
		resp.CuisineType = res.CuisineType
		resp.Address = res.Address
		resp.TimeToEat = res.TimeToEat
	}

	c.JSON(http.StatusOK, resp)
}