# backend/item-service/models/item.go
package models

import (
	"time"

	"github.com/google/uuid"
)

type Item struct {
	ID uint `gorm:"primaryKey;autoIncrement" json:"id"`
	// Generated by the frontend when the Add page opens, so the link parser
	// channel and the saved item share one id. Null for older items.
	UUID            *uuid.UUID     `gorm:"type:uuid;uniqueIndex" json:"uuid"`
	ItemName        string         `gorm:"size:255;not null" json:"item_name"`
	Quantity        int            `gorm:"default:1;not null" json:"quantity"`
	BuyURL          *string        `gorm:"size:2048" json:"buy_url"`
	ShopName        *string        `gorm:"size:255" json:"shop_name"`
	CreatedAt       time.Time      `gorm:"autoCreateTime" json:"created_at"`
	ItemType        string         `gorm:"size:50;not null" json:"item_type"`
	Bought          bool           `gorm:"default:false;not null" json:"bought"`
	ItemMediaFolder *string        `gorm:"size:4096" json:"item_media_folder"`
	Owner           string         `gorm:"size:255" json:"owner"`
	DeletedAt       *time.Time     `gorm:"default:null" json:"deleted_at,omitempty"`
	Metadata        map[string]any `gorm:"type:jsonb;serializer:json" json:"metadata,omitempty"`
}

type Clothes struct {
	ID    uint    `gorm:"primaryKey" json:"id"`
	Size  *string `gorm:"size:20" json:"size"`
	Color *string `gorm:"size:50" json:"color"`
	Brand *string `gorm:"size:100" json:"brand"`
}

type FoodAndDrink struct {
	ID       uint     `gorm:"primaryKey" json:"id"`
	Sugar    *string  `gorm:"size:100" json:"sugar"`
	Size     *string  `gorm:"size:20" json:"size"`
	Notes    *string  `gorm:"size:500" json:"notes"`
	Toppings []string `gorm:"serializer:json" json:"toppings"`
}

type Others struct {
	ID       uint    `gorm:"primaryKey" json:"id"`
	Category *string `gorm:"size:100" json:"category"`
	Notes    *string `gorm:"size:500" json:"notes"`
}

// Restaurant only needs a name up front; the rest can be filled in later
type Restaurant struct {
	ID          uint    `gorm:"primaryKey" json:"id"`
	MainFood    *string `gorm:"size:255" json:"main_food"`
	CuisineType *string `gorm:"size:100" json:"cuisine_type"`
	Address     *string `gorm:"size:500" json:"address"`
	TimeToEat   *string `gorm:"size:100" json:"time_to_eat"`
}

// AnyItem is used to serialize and deserialize the any item
type AnyItem struct {
	Item
	Size        *string        `json:"size,omitempty"`
	Color       *string        `json:"color,omitempty"`
	Brand       *string        `json:"brand,omitempty"`
	Sugar       *string        `json:"sugar,omitempty"`
	Notes       *string        `json:"notes,omitempty"`
	Toppings    []string       `json:"toppings,omitempty"`
	Category    *string        `json:"category,omitempty"`
	MainFood    *string        `json:"main_food,omitempty"`
	CuisineType *string        `json:"cuisine_type,omitempty"`
	Address     *string        `json:"address,omitempty"`
	TimeToEat   *string        `json:"time_to_eat,omitempty"`
	Additional  map[string]any `json:"additional,omitempty"`
}